pyproject.toml
[project]
name = "python-google-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arcadepy>=1.10.0",
    "google-adk>=1.22.1",
    "litellm>=1.81.0",
    "rich>=14.2.0",
]